requirements.txt
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dataclasses-json==0.6.5
deepdiff==7.0.1
emoji==2.11.1
filetype==1.2.0
idna==3.7
joblib==1.4.2
jsonpath-python==1.0.6
langdetect==1.0.9
lxml==5.2.1
marshmallow==3.21.2
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
ordered-set==4.1.0
packaging==24.0
pathspec==0.12.1
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
platformdirs==4.2.1
pycparser==2.22
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pypdf==4.2.0
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
python-iso639==2024.4.27
python-magic==0.4.27
rapidfuzz==3.9.0
regex==2024.4.28
requests==2.31.0
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
tomli==2.0.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
unstructured==0.13.6
unstructured-client==0.22.0
urllib3==2.2.1
wrapt==1.16.0
